+++ src/parser/peg/transformer/CMakeLists.txt
@@ -29,6 +29,7 @@ add_library_unity(
   transform_explain.cpp
   transform_export.cpp
   transform_expression.cpp
+  transform_generic_copy_option.cpp
   transform_import.cpp
   transform_insert.cpp
   transform_load.cpp
